refactor(api): move redis client
